Add readme and configuration to plugin panel
When you see the plugin in the plugin list (control panel) it will now
have two tabs. The simple readme (tells you what this plugin does) and a
panel for configuration.

Though it doesn't do a lot for in browser (well sort of) it does let you
adjust configuration while working in http://localhost:8080/ then use
that configuration during the renderTiddler part. Least editing the
config is easy and standard.