Rewrite/Refactor all the things
I wanted to customize the macro and realized there was some awkwardness
happening with the use of a manual string building. This came to light
when I thought a good exporter should be useable from the export menu.

This spawned a rewrite from scratch. I implemented a DomBuilder so make
the construction easy. It uses the built-in TiddlyWiki dom building
utils for this. Works in both Node and Browser.